requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-community==0.0.10
langchain-sambanova
pymongo==4.6.1
bcrypt==4.1.2
pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.25.2
plotly==5.17.0
streamlit-authenticator==0.2.3